requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
anyio==4.9.0
APScheduler==3.11.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.7.14
chardet==3.0.4
charset-normalizer==3.4.2
click==8.1.8
deep-translator==1.11.4
edge-tts==7.0.2
exceptiongroup==1.3.0
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.10.0
gTTS==2.5.4
h11==0.16.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.9
httpx==0.28.1
hyperframe==5.2.0
idna==2.10
Jinja2==3.1.6
llvmlite==0.43.0
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.6.3
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
openai-whisper==20250625
opencv-python==4.13.0.90
pillow==11.3.0
propcache==0.3.2
pypinyin==0.55.0
python-dotenv==1.1.1
python-telegram-bot==22.3
regex==2026.1.15
requests==2.32.4
rfc3986==1.5.0
sniffio==1.3.1
soupsieve==2.7
srt==3.5.3
sympy==1.14.0
tabulate==0.9.0
tiktoken==0.12.0
torch==2.8.0
tqdm==4.67.3
typing_extensions==4.14.1
tzlocal==5.3.1
urllib3==2.5.0
yarl==1.20.1
